backend/pyproject.toml
[project]
name = "secure-renewals-backend"
version = "0.1.0"
description = "Secure Renewals internal API"
authors = [{name = "Secure Renewals", email = "security@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.6.1",
]

[tool.uv]
package = true
